# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: URL Shortener Service API
  description: |
    Production-grade, high-concurrency URL Shortener service.

    Provides endpoints to create short URLs, resolve short codes to their
    original destinations, retrieve click analytics, and check service health.

    All error responses conform to **RFC 7807 - Problem Details for HTTP APIs**.
    Rate limiting is enforced per client and communicated via `X-RateLimit-*`
    response headers.
  version: 1.0.0
  contact:
    name: URL Shortener Platform Team
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.urlshortener.example.com
    description: Production
  - url: https://staging.api.urlshortener.example.com
    description: Staging
  - url: http://localhost:8080
    description: Local development
...
paths:
  /api/v1/urls:
    post:
      tags: [URLs]
      operationId: shortenUrl
      summary: Create a shortened URL
      description: |
        Creates a new short URL for the supplied `originalUrl`. Optionally accepts
        a `customAlias` (which must be unique) and an `expiresAt` timestamp after
        which the short code will no longer resolve.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUrlRequest'
            examples:
              minimal:
                summary: Minimal request
                value:
                  originalUrl: https://example.com/some/very/long/path?with=query
              withAlias:
                summary: Custom alias with expiration
                value:
                  originalUrl: https://example.com/campaign/launch
                  customAlias: launch-2026
                  expiresAt: '2026-12-31T23:59:59Z'
      responses:
        '201':
          description: Short URL created successfully.
          headers:
            Location:
              description: Absolute URL of the newly created short URL resource.
              required: true
              schema:
                type: string
                format: uri
                example: https://api.urlshortener.example.com/api/v1/urls/sK9xN4pQ
            X-RateLimit-Limit:
              $ref: '#/components/headers/XRateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/XRateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/XRateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortUrlResponse'
              examples:
                created:
                  value:
                    shortCode: sK9xN4pQ
                    shortUrl: https://api.urlshortener.example.com/sK9xN4pQ
                    originalUrl: https://example.com/some/very/long/path?with=query
                    createdAt: '2026-08-02T12:34:56Z'
                    expiresAt: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  headers:
    XRateLimitLimit:
      description: The maximum number of requests permitted within the current rate-limit window.
      schema:
        type: integer
        format: int32
        minimum: 1
        example: 1000
    XRateLimitRemaining:
      description: The number of requests remaining in the current rate-limit window.
      schema:
        type: integer
        format: int32
        minimum: 0
        example: 987
    XRateLimitReset:
      description: Unix epoch seconds indicating when the current rate-limit window resets.
      schema:
        type: integer
        format: int64
        example: 1754140800
  schemas:
    CreateUrlRequest:
      type: object
      required: [originalUrl]
      additionalProperties: false
      properties:
        originalUrl:
          type: string
          format: uri
          description: The absolute URL to be shortened. Must include an `http` or `https` scheme.
          minLength: 1
          maxLength: 2048
          pattern: '^https?://.+'
          example: https://example.com/some/very/long/path?with=query
        customAlias:
          type: string
          description: |
            Optional custom short code. If provided, this alias will be used
            instead of a system-generated code. Must be unique across the
            service.
          minLength: 4
          maxLength: 32
          pattern: '^[a-zA-Z0-9]{4,32}$'
          example: launch-2026
          nullable: true
        expiresAt:
          type: string
          format: date-time
          description: |
            Optional RFC 3339 timestamp (UTC) at which the short code will
            expire. Must be in the future. If omitted, the short URL does
            not expire.
          example: '2026-12-31T23:59:59Z'
          nullable: true
    ShortUrlResponse:
      type: object
      required: [shortCode, shortUrl, originalUrl, createdAt]
      properties:
        shortCode:
          type: string
          description: The generated (or supplied) short code.
          minLength: 4
          maxLength: 32
          pattern: '^[a-zA-Z0-9]{4,32}$'
          example: sK9xN4pQ
        shortUrl:
          type: string
          format: uri
          description: The fully qualified short URL to be shared with end users.
          example: https://api.urlshortener.example.com/sK9xN4pQ
        originalUrl:
          type: string
          format: uri
          description: The original long URL that the short code resolves to.
          example: https://example.com/some/very/long/path?with=query
        createdAt:
          type: string
          format: date-time
          description: RFC 3339 timestamp (UTC) when the short URL was created.
          example: '2026-08-02T12:34:56Z'
        expiresAt:
          type: string
          format: date-time
          description: RFC 3339 timestamp (UTC) at which the short code expires, if applicable.
          nullable: true
          example: '2026-12-31T23:59:59Z'
...
    ProblemDetail:
      type: object
      description: |
        RFC 7807 - Problem Details for HTTP APIs. Machine-readable error
        payload returned for all error responses.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type.
          default: about:blank
          example: https://api.urlshortener.example.com/problems/validation-error
        title:
          type: string
          description: A short, human-readable summary of the problem type.
          example: Bad Request
        status:
          type: integer
          format: int32
          description: The HTTP status code generated by the origin server.
          minimum: 400
          maximum: 599
          example: 400
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
          example: "'originalUrl' must be a valid absolute http(s) URL."
        instance:
          type: string
          format: uri-reference
          description: URI reference that identifies the specific occurrence of the problem.
          example: /api/v1/urls
        errors:
          type: array
          description: Optional list of field-level validation violations.
          items:
            $ref: '#/components/schemas/FieldViolation'
    FieldViolation:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          description: JSON pointer or field name that failed validation.
          example: originalUrl
        message:
          type: string
          description: Human-readable validation error message.
          example: must be a valid http(s) URL
        rejectedValue:
          description: The value that was rejected (may be omitted for sensitive fields).
          nullable: true
          example: not-a-url
  responses:
    BadRequest:
      description: The request payload failed validation.
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/XRateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/XRateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/XRateLimitReset'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            invalidUrl:
              value:
                type: https://api.urlshortener.example.com/problems/validation-error
                title: Bad Request
                status: 400
                detail: One or more request fields are invalid.
                instance: /api/v1/urls
                errors:
                  - field: originalUrl
                    message: must be a valid http(s) URL
                    rejectedValue: not-a-url
...
    Conflict:
      description: The supplied `customAlias` is already taken.
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/XRateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/XRateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/XRateLimitReset'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            aliasTaken:
              value:
                type: https://api.urlshortener.example.com/problems/alias-conflict
                title: Conflict
                status: 409
                detail: "The requested customAlias 'launch-2026' is already in use."
                instance: /api/v1/urls
    TooManyRequests:
      description: The caller has exceeded the configured rate limit.
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/XRateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/XRateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/XRateLimitReset'
        Retry-After:
          description: Seconds the client should wait before retrying.
          schema:
            type: integer
            format: int32
            minimum: 1
            example: 30
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            throttled:
              value:
                type: https://api.urlshortener.example.com/problems/rate-limit-exceeded
                title: Too Many Requests
                status: 429
                detail: Rate limit exceeded. Retry after the window resets.
                instance: /api/v1/urls
    InternalServerError:
      description: An unexpected error occurred while processing the request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            unexpected:
              value:
                type: https://api.urlshortener.example.com/problems/internal-error
                title: Internal Server Error
                status: 500
                detail: An unexpected error occurred. Please retry later.
                instance: /api/v1/urls